Adding authentication apis

+++ src/main/java/com/shubhu/staybooking/airBnbApp/service/HotelServiceImpl.java
@@ -5,12 +5,15 @@ import com.shubhu.staybooking.airBnbApp.dto.HotelInfoDto;
 import com.shubhu.staybooking.airBnbApp.dto.RoomDto;
 import com.shubhu.staybooking.airBnbApp.entity.Hotel;
 import com.shubhu.staybooking.airBnbApp.entity.Room;
+import com.shubhu.staybooking.airBnbApp.entity.User;
 import com.shubhu.staybooking.airBnbApp.exception.ResourceNotFoundException;
+import com.shubhu.staybooking.airBnbApp.exception.UnAuthorisedException;
 import com.shubhu.staybooking.airBnbApp.repository.HotelRepository;
 import com.shubhu.staybooking.airBnbApp.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
@@ -31,6 +34,8 @@ public class HotelServiceImpl implements HotelService {
         log.info("Creating a new hotel with name : {}", hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
         hotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with ID : {}", hotelDto.getId());
         return modelMapper.map(hotel, HotelDto.class);
@@ -42,6 +47,10 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID : " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This owner does not own this hotel with id : " + id);
+        }
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -51,6 +60,10 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID : " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This owner does not own this hotel with id : " + id);
+        }
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
@@ -63,7 +76,10 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID : " + id));
-
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This owner does not own this hotel with id : " + id);
+        }
         //Assuming only do it once
         for (Room room : hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
@@ -80,8 +96,11 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID : " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This owner does not own this hotel with id : " + id);
+        }
         hotel.setActive(true);
-
         //Assuming only do it once
         for (Room room : hotel.getRooms()) {
             inventoryService.initializeRoomForAYear(room);
